pyproject.toml
[project]
authors = [
  { name="Cedric Goubard", email="goubard.cedric@gmail.com" },
]
description = "Add a short description here"
name = "pupilzmq"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["opencv-python>=4.11.0.86,<5", "pupil-labs-realtime-api>=1.3.6,<2", "matplotlib>=3.10.0,<4"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pupilzmq"]


[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.pypi-dependencies]
pupilzmq = { path = ".", editable = true }

[tool.pixi.tasks]

[tool.pixi.dependencies]
pyyaml = ">=6.0.2,<7"
